Move ConsoleProvider to its own file
